+++ vpx_dsp/x86/inv_txfm_sse2.c
@@ -1006,8 +1006,7 @@ static INLINE void idct32_34_8x32_quarter_1(const __m128i *const in /*in[32]*/,
   __m128i step1[8], step2[8];
 
   // stage 3
-  butterfly(in[4], zero, (int)cospi_28_64, (int)cospi_4_64, &step1[4],
-            &step1[7]);
+  butterfly(in[4], zero, cospi_28_64, cospi_4_64, &step1[4], &step1[7]);
 
   // stage 4
   step2[0] = butterfly_cospi16(in[0]);
@@ -1022,8 +1021,7 @@ static INLINE void idct32_34_8x32_quarter_1(const __m128i *const in /*in[32]*/,
   step1[2] = step2[0];
   step1[3] = step2[0];
   step1[4] = step2[4];
-  butterfly(step2[6], step2[5], (int)cospi_16_64, (int)cospi_16_64, &step1[5],
-            &step1[6]);
+  butterfly(step2[6], step2[5], cospi_16_64, cospi_16_64, &step1[5], &step1[6]);
   step1[7] = step2[7];
 
   // stage 6
@@ -1046,10 +1044,8 @@ static INLINE void idct32_34_8x32_quarter_2(const __m128i *const in /*in[32]*/,
   __m128i step1[16], step2[16];
 
   // stage 2
-  butterfly(in[2], zero, (int)cospi_30_64, (int)cospi_2_64, &step2[8],
-            &step2[15]);
-  butterfly(zero, in[6], (int)cospi_6_64, (int)cospi_26_64, &step2[11],
-            &step2[12]);
+  butterfly(in[2], zero, cospi_30_64, cospi_2_64, &step2[8], &step2[15]);
+  butterfly(zero, in[6], cospi_6_64, cospi_26_64, &step2[11], &step2[12]);
 
   // stage 3
   step1[8] = step2[8];
@@ -1082,24 +1078,20 @@ static INLINE void idct32_34_8x32_quarter_3_4(
   __m128i step1[32];
 
   // stage 1
-  butterfly(in[1], zero, (int)cospi_31_64, (int)cospi_1_64, &step1[16],
-            &step1[31]);
-  butterfly(zero, in[7], (int)cospi_7_64, (int)cospi_25_64, &step1[19],
-            &step1[28]);
-  butterfly(in[5], zero, (int)cospi_27_64, (int)cospi_5_64, &step1[20],
-            &step1[27]);
-  butterfly(zero, in[3], (int)cospi_3_64, (int)cospi_29_64, &step1[23],
-            &step1[24]);
+  butterfly(in[1], zero, cospi_31_64, cospi_1_64, &step1[16], &step1[31]);
+  butterfly(zero, in[7], cospi_7_64, cospi_25_64, &step1[19], &step1[28]);
+  butterfly(in[5], zero, cospi_27_64, cospi_5_64, &step1[20], &step1[27]);
+  butterfly(zero, in[3], cospi_3_64, cospi_29_64, &step1[23], &step1[24]);
 
   // stage 3
-  butterfly(step1[31], step1[16], (int)cospi_28_64, (int)cospi_4_64, &step1[17],
+  butterfly(step1[31], step1[16], cospi_28_64, cospi_4_64, &step1[17],
             &step1[30]);
-  butterfly(step1[28], step1[19], -(int)cospi_4_64, (int)cospi_28_64,
-            &step1[18], &step1[29]);
-  butterfly(step1[27], step1[20], (int)cospi_12_64, (int)cospi_20_64,
-            &step1[21], &step1[26]);
-  butterfly(step1[24], step1[23], -(int)cospi_20_64, (int)cospi_12_64,
-            &step1[22], &step1[25]);
+  butterfly(step1[28], step1[19], -cospi_4_64, cospi_28_64, &step1[18],
+            &step1[29]);
+  butterfly(step1[27], step1[20], cospi_12_64, cospi_20_64, &step1[21],
+            &step1[26]);
+  butterfly(step1[24], step1[23], -cospi_20_64, cospi_12_64, &step1[22],
+            &step1[25]);
 
   idct32_8x32_quarter_3_4_stage_4_to_7(step1, out);
 }
@@ -1145,16 +1137,12 @@ static INLINE void idct32_1024_8x32_quarter_1(
   __m128i step1[8], step2[8];
 
   // stage 3
-  butterfly(in[4], in[28], (int)cospi_28_64, (int)cospi_4_64, &step1[4],
-            &step1[7]);
-  butterfly(in[20], in[12], (int)cospi_12_64, (int)cospi_20_64, &step1[5],
-            &step1[6]);
+  butterfly(in[4], in[28], cospi_28_64, cospi_4_64, &step1[4], &step1[7]);
+  butterfly(in[20], in[12], cospi_12_64, cospi_20_64, &step1[5], &step1[6]);
 
   // stage 4
-  butterfly(in[0], in[16], (int)cospi_16_64, (int)cospi_16_64, &step2[1],
-            &step2[0]);
-  butterfly(in[8], in[24], (int)cospi_24_64, (int)cospi_8_64, &step2[2],
-            &step2[3]);
+  butterfly(in[0], in[16], cospi_16_64, cospi_16_64, &step2[1], &step2[0]);
+  butterfly(in[8], in[24], cospi_24_64, cospi_8_64, &step2[2], &step2[3]);
   step2[4] = _mm_add_epi16(step1[4], step1[5]);
   step2[5] = _mm_sub_epi16(step1[4], step1[5]);
   step2[6] = _mm_sub_epi16(step1[7], step1[6]);
@@ -1166,8 +1154,7 @@ static INLINE void idct32_1024_8x32_quarter_1(
   step1[2] = _mm_sub_epi16(step2[1], step2[2]);
   step1[3] = _mm_sub_epi16(step2[0], step2[3]);
   step1[4] = step2[4];
-  butterfly(step2[6], step2[5], (int)cospi_16_64, (int)cospi_16_64, &step1[5],
-            &step1[6]);
+  butterfly(step2[6], step2[5], cospi_16_64, cospi_16_64, &step1[5], &step1[6]);
   step1[7] = step2[7];
 
   // stage 6
@@ -1189,14 +1176,10 @@ static INLINE void idct32_1024_8x32_quarter_2(
   __m128i step1[16], step2[16];
 
   // stage 2
-  butterfly(in[2], in[30], (int)cospi_30_64, (int)cospi_2_64, &step2[8],
-            &step2[15]);
-  butterfly(in[18], in[14], (int)cospi_14_64, (int)cospi_18_64, &step2[9],
-            &step2[14]);
-  butterfly(in[10], in[22], (int)cospi_22_64, (int)cospi_10_64, &step2[10],
-            &step2[13]);
-  butterfly(in[26], in[6], (int)cospi_6_64, (int)cospi_26_64, &step2[11],
-            &step2[12]);
+  butterfly(in[2], in[30], cospi_30_64, cospi_2_64, &step2[8], &step2[15]);
+  butterfly(in[18], in[14], cospi_14_64, cospi_18_64, &step2[9], &step2[14]);
+  butterfly(in[10], in[22], cospi_22_64, cospi_10_64, &step2[10], &step2[13]);
+  butterfly(in[26], in[6], cospi_6_64, cospi_26_64, &step2[11], &step2[12]);
 
   // stage 3
   step1[8] = _mm_add_epi16(step2[8], step2[9]);
@@ -1229,24 +1212,16 @@ static INLINE void idct32_1024_8x32_quarter_3_4(
   __m128i step1[32], step2[32];
 
   // stage 1
-  butterfly(in[1], in[31], (int)cospi_31_64, (int)cospi_1_64, &step1[16],
-            &step1[31]);
-  butterfly(in[17], in[15], (int)cospi_15_64, (int)cospi_17_64, &step1[17],
-            &step1[30]);
-  butterfly(in[9], in[23], (int)cospi_23_64, (int)cospi_9_64, &step1[18],
-            &step1[29]);
-  butterfly(in[25], in[7], (int)cospi_7_64, (int)cospi_25_64, &step1[19],
-            &step1[28]);
+  butterfly(in[1], in[31], cospi_31_64, cospi_1_64, &step1[16], &step1[31]);
+  butterfly(in[17], in[15], cospi_15_64, cospi_17_64, &step1[17], &step1[30]);
+  butterfly(in[9], in[23], cospi_23_64, cospi_9_64, &step1[18], &step1[29]);
+  butterfly(in[25], in[7], cospi_7_64, cospi_25_64, &step1[19], &step1[28]);
 
-  butterfly(in[5], in[27], (int)cospi_27_64, (int)cospi_5_64, &step1[20],
-            &step1[27]);
-  butterfly(in[21], in[11], (int)cospi_11_64, (int)cospi_21_64, &step1[21],
-            &step1[26]);
+  butterfly(in[5], in[27], cospi_27_64, cospi_5_64, &step1[20], &step1[27]);
+  butterfly(in[21], in[11], cospi_11_64, cospi_21_64, &step1[21], &step1[26]);
 
-  butterfly(in[13], in[19], (int)cospi_19_64, (int)cospi_13_64, &step1[22],
-            &step1[25]);
-  butterfly(in[29], in[3], (int)cospi_3_64, (int)cospi_29_64, &step1[23],
-            &step1[24]);
+  butterfly(in[13], in[19], cospi_19_64, cospi_13_64, &step1[22], &step1[25]);
+  butterfly(in[29], in[3], cospi_3_64, cospi_29_64, &step1[23], &step1[24]);
 
   // stage 2
   step2[16] = _mm_add_epi16(step1[16], step1[17]);
@@ -1270,16 +1245,16 @@ static INLINE void idct32_1024_8x32_quarter_3_4(
   // stage 3
   step1[16] = step2[16];
   step1[31] = step2[31];
-  butterfly(step2[30], step2[17], (int)cospi_28_64, (int)cospi_4_64, &step1[17],
+  butterfly(step2[30], step2[17], cospi_28_64, cospi_4_64, &step1[17],
             &step1[30]);
-  butterfly(step2[29], step2[18], -(int)cospi_4_64, (int)cospi_28_64,
-            &step1[18], &step1[29]);
+  butterfly(step2[29], step2[18], -cospi_4_64, cospi_28_64, &step1[18],
+            &step1[29]);
   step1[19] = step2[19];
   step1[20] = step2[20];
-  butterfly(step2[26], step2[21], (int)cospi_12_64, (int)cospi_20_64,
-            &step1[21], &step1[26]);
-  butterfly(step2[25], step2[22], -(int)cospi_20_64, (int)cospi_12_64,
-            &step1[22], &step1[25]);
+  butterfly(step2[26], step2[21], cospi_12_64, cospi_20_64, &step1[21],
+            &step1[26]);
+  butterfly(step2[25], step2[22], -cospi_20_64, cospi_12_64, &step1[22],
+            &step1[25]);
   step1[23] = step2[23];
   step1[24] = step2[24];
   step1[27] = step2[27];
